requirements.txt
flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
sqlite3==2.6.0
pytz==2023.3
